Refactor: move ZeroMQ interface into network module
Issue #596 の意図（network層を集約）に合わせ、C++のZeroMQ通信コンポーネントを
src/network と include/network に移動し、CMake と include 参照を更新する。
