+++ src/network/zeromq_interface.cpp
@@ -1,4 +1,4 @@
-#include "zeromq_interface.h"
+#include "network/zeromq_interface.h"
 
 #include <iostream>
 #include <map>
